+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.0.16] - 2020-10-15
+### [1.0.17] - 2020-10-18
+
+Allowing configuration of `val` properties.
+Added `KofikoSettings.configureReadonlyProperties` option.
+
+### [1.0.16] - 2020-10-15
 
 Initial official release.